requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
multidict==6.0.4
networkx==3.1
openai==0.28.1
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.8
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.21
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
yarl==1.9.2
